Match behavior of Ruby implementation for embedded null values.
The Ruby implementation parses the null values in JSON to nil values
in Ruby which then end up as empty string when put into a format
string.

I'm not sure if our query transformation provides for a way to
generate a valid query. Arguably indexing the null value using the
string "null" might make for a better setup.

Conflicts:
	test/chef_index_expand_tests.erl